requirements.txt
langchain_community
langchain_core
streamlit
langchain_groq
python-dotenv
PyPDF
faiss-cpu
langchain_huggingface
langchain_chroma
wikipedia
ipykernel
arxiv
duckduckgo-search
validators
youtube_transcript_api
pytube
unstructured
numexpr
huggingface_hub